combine cogs, display_emoji, docstring
