+++ utils/bots/ModBot/cogs/mod.py
@@ -3,18 +3,23 @@ import math
 import peewee
 import discord
 from core import common, database
-from core.common import hexColors
+from core.common import hexColors, Emoji
 from discord.ext import commands
 
 
 class PunishmentTag(commands.Cog):
-    """Commands related to our dynamic tag system."""
+    """Moderation commands"""
 
     def __init__(self, bot):
         self.bot = bot
+        self.__cog_name__ = "Moderation"
+
+    @property
+    def display_emoji(self) -> discord.PartialEmoji:
+        return discord.PartialEmoji(name="modshield", id=957316915158728827)
 
     @staticmethod
-    def get_by_index(self, index):
+    def get_by_index(index):
         for i, t in enumerate(database.PunishmentTag.select()):
             if i + 1 == index:
                 return t
@@ -123,6 +128,95 @@ class PunishmentTag(commands.Cog):
         embed = await common.paginate_embed(
             self.bot, ctx, embed, populate_embed, get_end(10), page=page
         )
+
+    @commands.command(aliases=["find"])
+    @commands.has_any_role(
+        "Moderator", "Mod", "Senior Mod", "Head Mod", 844013914609680384, "HR"
+    )
+    async def info(
+            self, ctx: commands.Context, user: commands.Greedy[discord.User] = []
+    ):
+        for user in user:
+            user: discord.User = user
+
+            value = None
+            typeval = None
+            banreason = None
+
+            embed = discord.Embed(
+                title="Queued Query",
+                description=f"I have started a new query for {user.display_name}",
+                color=discord.Color.gold(),
+            )
+            embed.set_footer(text="This may take a moment.")
+            msg = await ctx.send(embed=embed)
+
+            member = ctx.guild.get_member(user.id)
+            if member is None:
+                banEntry = await ctx.guild.fetch_ban(user)
+
+                if banEntry is not None:
+                    value = Emoji.deny
+                    typeval = "Banned"
+                    banreason = banEntry.reason
+
+                else:
+                    value = Emoji.question
+                    typeval = "Not in the Server"
+
+            else:
+                value = Emoji.confirm
+                typeval = "In the Server"
+
+            if banreason is None:
+                embed = discord.Embed(
+                    description=f"`ID: {user.id}` | {user.mention} found with the nickname: **{user.display_name}**\u0020",
+                    color=discord.Color.green(),
+                )
+                embed.set_author(
+                    name={user.name}, icon_url=user.avatar.url, url=user.avatar.url
+                )
+                embed.add_field(
+                    name="Membership Status", value=f"\u0020{value} `{typeval}`"
+                )
+
+            else:
+                embed = discord.Embed(
+                    description=f"`ID: {user.id}` | {user.mention} found with the nickname: {user.display_name}\u0020",
+                    color=discord.Color.green(),
+                )
+                embed.set_author(
+                    name={user.name}, icon_url=user.avatar.url, url=user.avatar.url
+                )
+                embed.add_field(
+                    name="Membership Status",
+                    value=f"\u0020{value} `{typeval}`\n{Emoji.space}{Emoji.barrow}**Ban Reason:** {banreason}",
+                )
+
+            await msg.edit(embed=embed)
+
+    @info.error
+    async def info_error(self, ctx, error):
+        if isinstance(error, (commands.UserNotFound, commands.errors.UserNotFound)):
+            embed = discord.Embed(
+                title="User Not Found",
+                description="Try using an actual user next time? :(",
+                color=discord.Color.red(),
+            )
+            await ctx.send(embed=embed)
+
+        elif isinstance(
+                error,
+                (commands.MissingRequiredArgument, commands.errors.MissingRequiredArgument),
+        ):
+            embed = discord.Embed(
+                title="User Not Found",
+                description="Try using an actual user next time? :(",
+                color=discord.Color.red(),
+            )
+            await ctx.send(embed=embed)
+        else:
+            raise error
 
 
 async def setup(bot):
